+++ iree/turbine/kernel/_support/tracing.py
@@ -1,24 +1,34 @@
+from abc import ABC, abstractmethod
+from typing import (
+    Optional,
+    TypeVar,
+    Callable,
+    Type,
+    cast,
+    Dict,
+    Tuple,
+)
+from types import FunctionType
+
+from ..compiler.ir import Operation
+
 import functools
 import warnings
-from abc import ABC, abstractmethod
-from types import FunctionType
-from typing import (
-    Callable,
-    Dict,
-    Optional,
-    Tuple,
-    Type,
-    TypeVar,
-    cast,
-)
 
-import sympy
 import torch.fx as fx
 
-from .. import ops
-from ..compiler.ir import Operation
-from ..lang.grid import Grid
+from .indexing import (
+    backed_sym_index_type,
+    BoundedRelation,
+    IndexExpr,
+    IndexSymbol,
+    IndexingContext,
+)
+import sympy
+
 from ..lang.kernel_buffer import KernelBuffer, KernelBufferMeta
+from ..lang.grid import Grid
+
 from ..lang.types import (
     Index,
 )
@@ -28,20 +38,12 @@ from ..ops.wave_ops import CustomOp, Placeholder, Iterate, Unknown
 from .regions import RegionGraph, SubgraphTracer
 
 from .. import ops
-
 from ..ops.base import (
     OpDispatcher,
 )
-from ..ops.wave_ops import CustomOp
+
 from . import context
 from .dtype import DataType
-from .indexing import (
-    BoundedRelation,
-    IndexingContext,
-    IndexSymbol,
-    backed_sym_index_type,
-)
-from .regions import RegionGraph, SubgraphTracer
 
 try:
     from typing import assert_type
@@ -335,17 +337,6 @@ class CompiledContext(BaseContext):
             "call_function",
             target=op,
             args=(shape, dtype, value),
-            kwargs={},
-        )
-
-    ### ========================================================================
-    ### Scan Operations
-    ### ========================================================================
-    def handle_vector_cumsum(self, op, vector, axis=None, acc=None):
-        return self.region_graph.create_proxy(
-            "call_function",
-            target=op,
-            args=(vector, axis, acc),
             kwargs={},
         )
