+++ iree/turbine/kernel/wave/wave.py
@@ -63,7 +63,7 @@ from .shared_memory_indexing import (
 )
 
 # Utils
-from .utils.symbol_utils import delinearize_index, subs_idxc, safe_subs
+from .utils.symbol_utils import subs_idxc, safe_subs
 from .utils.classes import KernelLaunchInfo
 from .utils.print_utils import print_trace, try_apply_pass
 from .utils.graph_utils import (
