+++ tests/kernel/wave/common/shapes.py
@@ -76,12 +76,6 @@ _e2e_test_shapes["gqa_bshd_decode_attention"] = [
     ),
 ]
 
-_e2e_test_shapes["test_block_reduce"] = [
-    (1, 256),
-    (256, 256),
-    (256, 1024),
-]
-
 test_names = [
     "attention",
     "bhsd_attention",
@@ -93,7 +87,6 @@ test_names = [
     "evoformer",
     "gqa_bshd_attention",
     "gqa_bshd_decode_attention",
-    "test_block_reduce",
 ]
 for test in test_names:
     _perf_test_shapes[test] = _e2e_test_shapes[test]
